# qgemm: optimize avxvnni QGEMM inner kernel for M=1 