@r-devulap 2024-11-26
Posting raw qgemm benchmark numbers for clarity: 

Before: 
```
-------------------------------------------------------------------------------------------------------------
Benchmark                                                                   Time             CPU   Iterations
-------------------------------------------------------------------------------------------------------------
QGEMM/UnsignedAPackB/M:1/N:512/K:512/Batch:1/Threads:1/real_time         4330 ns         4330 ns       161969
QGEMM/UnsignedAPackB/M:1/N:512/K:1024/Batch:1/Threads:1/real_time        9027 ns         9027 ns        77210
QGEMM/UnsignedAPackB/M:1/N:1024/K:1024/Batch:1/Threads:1/real_time      17867 ns        17866 ns        39329
-------------------------------------------------------------------------------------------------------------
```

After: 
```
-------------------------------------------------------------------------------------------------------------
Benchmark                                                                   Time             CPU   Iterations
-------------------------------------------------------------------------------------------------------------
QGEMM/UnsignedAPackB/M:1/N:512/K:512/Batch:1/Threads:1/real_time         3137 ns         3136 ns       221932
QGEMM/UnsignedAPackB/M:1/N:512/K:1024/Batch:1/Threads:1/real_time        6385 ns         6385 ns       109727
QGEMM/UnsignedAPackB/M:1/N:1024/K:1024/Batch:1/Threads:1/real_time      12499 ns        12498 ns        55934
```