@liqunfu 2024-12-11
will avx2 benefits from the same - given redundant registers in RowCount=1 cases?