@r-devulap 2024-12-15
> will avx2 benefits from the same - given redundant registers in RowCount=1 cases?

Not sure. It does have the extra spare registers but the AVX2 code path runs on CPU prior to Icelake (2019), so the CPU is much older technology and not sure if it will show the same benefits as the newer ones.