@liqunfu 2025-01-09
> > will avx2 benefits from the same - given redundant registers in RowCount=1 cases?
> 
> Not sure. It does have the extra spare registers but the AVX2 code path runs on CPU prior to Icelake (2019), so the CPU is much older technology and not sure if it will show the same benefits as the newer ones.

I am worried for the extra code path just for M1 VNNI, I am thinking if avx2 will also benefit for the extra registers there will be only one code path. That makes code more maintainable.