@liqunfu 2025-02-06
> > I am worried for the extra code path just for M1 VNNI, I am thinking if avx2 will also benefit from the extra registers there will be only one code path. That makes code more maintainable.
> 
> I don't think this code is hard to maintain and I have reservations about spending time and effort in making this work for AVX2 for 2 reasons:
> 
> (1) Processors that exercise AVX2 code path are nearly 9 yrs old (Skylake generation). An average laptop/desktop lifespan is lesser than that.
> 
> (2) Even if we were to exercise the AVX2 code path with this special case for M = 1, I doubt we will see the performance benefits. Skylake generation has just 2 load ports (P2 and P3) and is very likely memory bound in that loop.

thank for the explanation!